Finished third task

+++ hospital-equipment/src/main/java/com/ISAproject/hospitalequipment/domain/Equipment.java
@@ -28,7 +28,7 @@ public class Equipment {
 
     public Double grade;
 
-    public Integer amount;
+    public Long amount;
 
 
     @JsonIgnore
@@ -36,7 +36,7 @@ public class Equipment {
     @OneToMany(mappedBy = "equipment", fetch=FetchType.EAGER, cascade = CascadeType.ALL)
 
     private Set<EquipmentStock> equipmentStocks;
-    public Equipment(Long id,String name, String description, Double grade, Integer amount){
+    public Equipment(Long id,String name, String description, Double grade, Long amount){
         this.id = id;
         this.name = name;
         this.description = description;
